% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{case_compare7}
\alias{case_compare7}
\title{Comparison function with data}
\usage{
case_compare7(state, observed, pars = NULL)
}
\arguments{
\item{state}{}

\item{observed}{}

\item{pars}{}
}
\value{
real
}
\description{
Default compare function (log likelihood)
}
\details{
See mcstate docs
}
\author{
Pete Dodd
}
