% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.particlefilter.R
\name{create.particlefilter}
\alias{create.particlefilter}
\title{Create a particle filter}
\usage{
create.particlefilter(
  data,
  comparefun,
  n_particles,
  n_threads = 4,
  seed = 5L,
  ...
)
}
\arguments{
\item{data}{must be mcstate particle filter data}

\item{comparefun}{mcstate compare function (log likelihood)}

\item{n_particles}{}

\item{n_threads}{}

\item{seed}{}

\item{...}{}
}
\value{
mcstate particle filter
}
\description{
This creates an object for running model as particle filtering etc.
}
\details{
See mcstate docs
}
\author{
Pete Dodd
}
