% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{extract.pops.multi}
\alias{extract.pops.multi}
\title{Postprocess model outputs for multiple chains}
\usage{
extract.pops.multi(Y, n_chain_steps = 100, out_type = "N")
}
\arguments{
\item{Y}{}

\item{n_chain_steps}{}

\item{out_type}{}
}
\value{
data.table
}
\description{
Collect total populations as data table to be used by each plotter
}
\details{
TODO check chains or particles?
}
\author{
Pete Dodd
}
