% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{extract.pops}
\alias{extract.pops}
\title{Postprocess model output}
\usage{
extract.pops(Y, chain_step_num = 1, out_type = "N")
}
\arguments{
\item{Y}{}

\item{chain_step_num}{}

\item{out_type}{}
}
\value{
data.table
}
\description{
Collect total populations as data table to be used by each plotter
}
\details{
TODO
}
\author{
Pete Dodd
}
