% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.parms.R
\name{get_TBI_prev}
\alias{get_TBI_prev}
\title{TB infection prevalence from model output}
\usage{
get_TBI_prev(X, time, grp = "adult")
}
\arguments{
\item{X}{Model output array, in same format as output of 'stocm()'}

\item{time}{which step}

\item{grp}{either 'adult', 'kid', or 'all' - which group to calculate for}
}
\value{
vector with mean and 95% quantile over particles for TBI prevalence at given timepoint
}
\description{
Calculate TBI prevalence at given timepoint from model output
}
\author{
Pete Dodd
}
