% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{HIV_prev_TB}
\alias{HIV_prev_TB}
\title{Plot HIV prevalence among TB notifications#'}
\usage{
HIV_prev_TB(Y, n_chain_steps = 100, separate = FALSE)
}
\arguments{
\item{Y}{input data}

\item{separate}{separate HIV by ART status}

\item{chain_step_num}{chain step}
}
\value{
ggplot2
}
\description{
Plot HIV prevalence among TB notifications#'
}
