% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{in_args}
\alias{in_args}
\title{Model Parameters}
\format{
## `parms`
A list in the right format for passing to `runmodel`
}
\usage{
in_args
}
\description{
A default set of parameter values for running the model
}
\keyword{datasets}
