% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot_compare_demog}
\alias{plot_compare_demog}
\title{Plot comparing total populations}
\usage{
plot_compare_demog(
  Y,
  n_chain_steps = 100,
  start_year = 2015,
  years = 6,
  eps = 0.25,
  by_comp = "none"
)
}
\arguments{
\item{Y}{input data}

\item{n_chain_steps}{chain step}

\item{start_year}{start year}

\item{years}{number of years}

\item{eps}{UI width}

\item{by_comp}{plot type}
}
\value{
ggplot2
}
\description{
Default median N per age group per year (not by patch)
}
\details{
TODO
}
\author{
Pete Dodd
}
