% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot_compare_noterate_agrgt}
\alias{plot_compare_noterate_agrgt}
\title{Plot average notification rates & compare with real data}
\usage{
plot_compare_noterate_agrgt(
  Y,
  n_chain_steps = 100,
  agrgt_by = c("patch"),
  realdata = TRUE,
  start_year = 2015
)
}
\arguments{
\item{Y}{input data}

\item{n_chain_steps}{chain steps}

\item{agrgt_by}{aggregate by}

\item{realdata}{logical (default TRUE) on whether to show real data points}

\item{start_year}{start for scale (default 2015)}
}
\value{
ggplot2
}
\description{
Plot average notification rates & compare with real data
}
