% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot_HIV_dynamic}
\alias{plot_HIV_dynamic}
\title{HIV by time}
\usage{
plot_HIV_dynamic(
  Y,
  start_year,
  chain_step_num = 1,
  out_type = "N",
  by_age = FALSE
)
}
\arguments{
\item{Y}{input data}

\item{start_year}{start year}

\item{chain_step_num}{chain step}

\item{out_type}{output type}

\item{by_age}{by age or not}
}
\value{
ggplot
}
\description{
HIV/ART over time
}
\details{
TODO
}
\author{
Pete Dodd
}
