% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot_HIV_in_TB}
\alias{plot_HIV_in_TB}
\title{PLotter for HIV in TB}
\usage{
plot_HIV_in_TB(X, start_year = 2015)
}
\arguments{
\item{X}{output from `run.model`}

\item{start_year}{the start year for the simulation}
}
\value{
a ggplot2 plot
}
\description{
Compare HIV in TB
}
\details{
This processes model outputs to compare the proportion of TB that is HIV+ over time
}
\author{
Pete Dodd
}
