% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot_pops_dynamic}
\alias{plot_pops_dynamic}
\title{Plot selected populations/compartments over time (absolute numbers, no comparison w/real data)}
\usage{
plot_pops_dynamic(
  Y,
  chain_step_num = 1,
  out_type = "N",
  start_year = 2015,
  by_comp = NULL
)
}
\arguments{
\item{Y}{}

\item{chain_step_num}{}

\item{out_type}{}

\item{by_comp}{}
}
\value{
ggplot2
}
\description{
Plot selected populations/compartments over time (absolute numbers, no comparison w/real data)
}
