% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot_TB_dynamics}
\alias{plot_TB_dynamics}
\title{Plot sample TB incidence}
\usage{
plot_TB_dynamics(
  Y,
  chain_step_num = 1,
  out_type = "incidence",
  separate = FALSE,
  by_age = F
)
}
\arguments{
\item{Y}{}

\item{chain_step_num}{}

\item{out_type}{}

\item{separate}{}

\item{by_age}{}
}
\value{
ggplot2
}
\description{
Plot sample TB incidence
}
