% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.parms.R
\name{restart_parms}
\alias{restart_parms}
\title{Restart parameters at given step}
\usage{
restart_parms(parms, restart_step, end_state, sim_length = NULL)
}
\arguments{
\item{parms}{Parameter list from previous runs, see 'get.parms()'}

\item{restart_step}{Step to restart at, in same units as 'tt' in parms}

\item{end_state}{Output from previous runs, in same format as output of 'stocm()' - should contain all timesteps up to and including restart_step}

\item{sim_length}{How long is the new simulation? If NULL, will use length of 'tt' in parms minus restart_step}
}
\value{
Updated parameter list with initial conditions and time-varying parameters updated to restart at given step
}
\description{
Update parameters for given step using results of previous runs
}
\details{
TODO
}
\author{
Pete Dodd
}
