% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.model.R
\name{run.CF}
\alias{run.CF}
\title{run counterfactuals}
\usage{
run.CF(
  parm,
  cfparm,
  times,
  n.particles = 1,
  pars_multi = FALSE,
  convert = FALSE
)
}
\arguments{
\item{parm}{base case parameters}

\item{cfparm}{counterfactual parameters}

\item{times}{times to run over}

\item{n.particles}{number of particles}

\item{pars_multi}{TODO}

\item{convert}{to dataframe?}
}
\value{
array or data.frame
}
\description{
For simulating counterfactuals
}
\details{
TODO
}
\author{
Pete Dodd
}
