% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.pmcmc.R
\name{run.pmcmc}
\alias{run.pmcmc}
\title{Run PMCMC inference on particle filter}
\usage{
run.pmcmc(
  particle.filter,
  parms,
  prior.list = NULL,
  initial.proposal.matrix = NULL,
  n.steps,
  n.burnin,
  n.chains,
  n.threads = 4,
  n.epochs = 1,
  save_restart = NULL,
  transform = NULL,
  mcmc_pars = NULL,
  returnall = FALSE
)
}
\arguments{
\item{particle.filter}{an mcstate particle filter object}

\item{parms}{base parameters list}

\item{prior.list}{priors for inference parameters}

\item{initial.proposal.matrix}{starting point for PMCMC proposal matric}

\item{n.steps}{number of PMCMC iterations}

\item{n.burnin}{number of interations to discard}

\item{n.chains}{number of PMCMC chains to run}

\item{n.threads}{number of threads to use (not tested)}

\item{n.epochs}{number of times to re-run PMCMC updating the proposal matrix based on previous run}

\item{save_restart}{where to save the restart}

\item{transform}{if not supplied will default to transform = function(theta) c(parms, as.list(theta))}

\item{mcmc_pars}{if not supplied, will create from prior.list & initial.proposal.matrix (& optionally transform)}

\item{returnall}{whether to return a list of things or just processed chains}
}
\value{
mcstate processed chains object (unless returnall=TRUE, in which case a list)
}
\description{
This is the main function for performing particle filter inference
}
\details{
TODO
}
\author{
Pete Dodd
}
