% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.parms.R
\name{get.parms}
\alias{get.parms}
\title{Create parameter object}
\usage{
get.parms(
  start_year,
  years,
  Dinit,
  ari0,
  hivoffset = 0,
  hivfac = 2,
  hivdecline = 0,
  hiv_init_override = -1,
  ART_haz = 0.5,
  ART_init_override = -1,
  hiv_checking = FALSE
)
}
\arguments{
\item{start_year}{Earliest year of simulation}

\item{years}{N years in simulation}

\item{Dinit}{Matrix of initial prevalences}

\item{ari0}{initial ARIs for initial state}

\item{hivoffset}{How many years ahead is Blantyre HIV incidence than MWI?}

\item{hivfac}{HIV incidence in Blantyre relative to MWI}

\item{hivdecline}{HIV decline rate from peak: overrides if >0}

\item{hiv_init_override}{if >0 provides initial HIV prevalence}

\item{ART_haz}{provides an ART initiation hazard}

\item{ART_init_override}{if >0 provides initial ART prevalence}

\item{hiv_checking}{if TRUE, prints HIV-related parameters and plots HIV incidence and ART initiation over time}
}
\value{
list
}
\description{
Create parameter object for model
}
\details{
Defaults for various things needed for model to run
}
\author{
Pete Dodd
}
