% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.parms.R
\name{get.parms}
\alias{get.parms}
\title{Create parameter object}
\usage{
get.parms(start_year, years)
}
\arguments{
\item{start_year}{TODO}

\item{years}{TODO}
}
\value{
list
}
\description{
Create parameter object for model
}
\details{
TODO
}
\author{
Pete Dodd
}
