% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.parms.R
\name{get.parms}
\alias{get.parms}
\title{Create parameter object}
\usage{
get.parms(start_year, years, Dinit, ari0, hivoffset = 5)
}
\arguments{
\item{start_year}{Earliest year of simulation}

\item{years}{N years in simulation}

\item{Dinit}{Matrix of initial prevalences}

\item{ari0}{initial ARIs for initial state}

\item{hivoffset}{How many years ahead is Blantyre HIV incidence than MWI?}
}
\value{
list
}
\description{
Create parameter object for model
}
\details{
TODO
}
\author{
Pete Dodd
}
