% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{HIV_prev_TB}
\alias{HIV_prev_TB}
\title{Plot HIV prevalence among TB notifications#'}
\usage{
HIV_prev_TB(Y, chain_step_num = 1, separate = FALSE, prev = "pc")
}
\arguments{
\item{Y}{input data}

\item{chain_step_num}{chain step}

\item{separate}{separate plots?}

\item{prev}{prevalence type?}
}
\value{
ggplot2
}
\description{
Plot HIV prevalence among TB notifications#'
}
