% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot_HIV_snapshot}
\alias{plot_HIV_snapshot}
\title{HIV snapshot}
\usage{
plot_HIV_snapshot(Y, chain_step_num = 1, out_type = "N", timestep = NULL)
}
\arguments{
\item{Y}{input data}

\item{chain_step_num}{chain step}

\item{out_type}{output type}

\item{timestep}{timestep for snapshot}
}
\value{
ggplot
}
\description{
HIV snapshot by age at single time
}
\details{
TODO
}
\author{
Pete Dodd
}
