% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot_TB_dynamics}
\alias{plot_TB_dynamics}
\title{Plot sample TB incidence}
\usage{
plot_TB_dynamics(
  Y,
  chain_step_num = 1,
  out_type = "incidence",
  separate = FALSE,
  by_age = FALSE,
  by_HIV = FALSE,
  wrap = TRUE,
  start_year = 2015
)
}
\arguments{
\item{Y}{input data}

\item{chain_step_num}{chain step number}

\item{out_type}{output type}

\item{separate}{separate plot?}

\item{by_age}{by age? (default FALSE)}

\item{wrap}{display with facet wrapping (default TRUE) or grids}

\item{start_year}{start for scale (default 2015)}

\item{by_hiv}{by HIV? (default FALSE)}
}
\value{
ggplot2
}
\description{
Plot sample TB incidence
}
