% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.model.R
\name{run.model}
\alias{run.model}
\title{Run dust version of model}
\usage{
run.model(parm, times, n.particles = 1, convert = FALSE)
}
\arguments{
\item{parm}{}

\item{times}{}

\item{n.particles}{}

\item{convert}{}
}
\value{
array
}
\description{
This will run the stochastic version of the model without reference to data
}
\details{
TODO
}
\author{
Pete Dodd
}
