% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.pmcmc.R
\name{run.pmcmc}
\alias{run.pmcmc}
\title{Run PMCMC inference on particle filter}
\usage{
run.pmcmc(
  particle.filter,
  parms,
  prior.list,
  initial.proposal.matrix,
  n.steps,
  n.burnin,
  n.chains,
  n.threads = 4,
  n.epochs = 1
)
}
\arguments{
\item{particle.filter}{TODO}

\item{parms}{TODO}

\item{prior.list}{TODO}

\item{initial.proposal.matrix}{TODO}

\item{n.steps}{TODO}

\item{n.burnin}{TODO}

\item{n.chains}{TODO}

\item{n.threads}{TODO}

\item{n.epochs}{TODO}
}
\value{
mcstate processed chains object
}
\description{
This is the main function for performing particle filter inference
}
\details{
TODO
}
\author{
Pete Dodd
}
