% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.pmcmc.R
\name{run.pmcmc}
\alias{run.pmcmc}
\title{Run PMCMC inference on particle filter}
\usage{
run.pmcmc(
  particle.filter,
  parms,
  prior.list,
  initial.proposal.matrix,
  n.steps,
  n.burnin,
  n.chains,
  n.threads = 4,
  n.epochs = 1
)
}
\arguments{
\item{particle.filter}{}

\item{parms}{}

\item{prior.list}{}

\item{initial.proposal.matrix}{}

\item{n.steps}{}

\item{n.burnin}{}

\item{n.chains}{}

\item{n.threads}{}

\item{n.epochs}{}
}
\value{
mcstate processed chains object
}
\description{
This is the main function for performing particle filter inference
}
\details{
TODO
}
\author{
Pete Dodd
}
